requirements.txt
openai-agents
python-dotenv
pydantic
arxiv
pymupdf
fastmcp[tasks]
uvicorn
fastapi
ngrok